backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
pydantic<3
pytesseract
pillow
google-genai
pandas
python-dotenv
